refactor: move all hanlders and handler tests into a api package
