+++ internal/transport/http/api/auth.go
@@ -1,4 +1,4 @@
-package auth
+package api
 
 import (
 	"errors"
@@ -10,17 +10,16 @@ import (
 	"github.com/golang-jwt/jwt/v5"
 
 	"url-short/internal/database"
-	"url-short/internal/transport/http/helper"
 )
 
-type handler struct {
+type authHandler struct {
 	// temporary now to allow transport, service and repository layers to be decoupled
 	database  *database.Queries
 	JWTSecret string
 }
 
-func NewAuthHandler(database *database.Queries, JWTSecret string) *handler {
-	return &handler{
+func NewAuthHandler(database *database.Queries, JWTSecret string) *authHandler {
+	return &authHandler{
 		database:  database,
 		JWTSecret: JWTSecret,
 	}
@@ -48,12 +47,12 @@ func ExtractAuthTokenFromRequest(r *http.Request) (string, error) {
 
 type authedHandeler func(http.ResponseWriter, *http.Request, database.User)
 
-func (handler *handler) AuthenticationMiddleware(nextHandler authedHandeler) http.HandlerFunc {
+func (handler *authHandler) AuthenticationMiddleware(nextHandler authedHandeler) http.HandlerFunc {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		requestToken, err := ExtractAuthTokenFromRequest(r)
 
 		if err != nil {
-			helper.RespondWithError(w, http.StatusUnauthorized, err.Error())
+			respondWithError(w, http.StatusUnauthorized, err.Error())
 			return
 		}
 
@@ -66,40 +65,40 @@ func (handler *handler) AuthenticationMiddleware(nextHandler authedHandeler) htt
 		)
 
 		if err != nil {
-			helper.RespondWithError(w, http.StatusUnauthorized, "invalid jwt")
+			respondWithError(w, http.StatusUnauthorized, "invalid jwt")
 			return
 		}
 
 		issuer, err := token.Claims.GetIssuer()
 
 		if err != nil {
-			helper.RespondWithError(w, http.StatusUnauthorized, "invalid jwt issuer")
+			respondWithError(w, http.StatusUnauthorized, "invalid jwt issuer")
 			return
 		}
 
 		if issuer != "url-short-auth" {
-			helper.RespondWithError(w, http.StatusUnauthorized, "invalid jwt issuer")
+			respondWithError(w, http.StatusUnauthorized, "invalid jwt issuer")
 			return
 		}
 
 		userID, err := token.Claims.GetSubject()
 
 		if err != nil {
-			helper.RespondWithError(w, http.StatusUnauthorized, "could not get subject from jwt")
+			respondWithError(w, http.StatusUnauthorized, "could not get subject from jwt")
 			return
 		}
 
 		userIDStr, err := strconv.Atoi(userID)
 
 		if err != nil {
-			helper.RespondWithError(w, http.StatusUnauthorized, "invalid jwt subject")
+			respondWithError(w, http.StatusUnauthorized, "invalid jwt subject")
 		}
 
 		user, err := handler.database.SelectUserByID(r.Context(), int32(userIDStr))
 
 		if err != nil {
 			log.Println(err)
-			helper.RespondWithError(w, http.StatusUnauthorized, "could not find user in database")
+			respondWithError(w, http.StatusUnauthorized, "could not find user in database")
 		}
 
 		nextHandler(w, r, user)
