--- internal/transport/http/users/main_test.go
@@ -1,1 +1,0 @@
-package users
